fix return default type

+++ tests/tier1/test_power.cpp
@@ -47,12 +47,12 @@ TEST_P(TestPower, execute_w_float)
 TEST_P(TestPower, execute_w_int)
 {
 
-  std::array<int32_t, 10 * 5 * 3> output;
+  std::array<float, 10 * 5 * 3> output;
   std::array<int32_t, 10 * 5 * 3> input;
-  std::array<int32_t, 10 * 5 * 3> valid;
+  std::array<float, 10 * 5 * 3> valid;
 
   std::fill(input.begin(), input.end(), static_cast<int32_t>(value));
-  std::fill(valid.begin(), valid.end(), static_cast<int32_t>(std::pow(value, exp)));
+  std::fill(valid.begin(), valid.end(), static_cast<float>(std::pow(value, exp)));
 
   auto gpu_input = cle::Array::create(10, 5, 3, 3, cle::dType::INT32, cle::mType::BUFFER, device);
   gpu_input->writeFrom(input.data());
